requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.3.0
pydantic-settings>=2.0.3
python-dotenv>=1.0.1
python-multipart>=0.0.6
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
openai>=1.0.0
PyPDF2>=3.0.0
numpy>=1.24.0
requests>=2.31.0
streamlit>=1.35.0
